Refactor: add folders to Authentication
