+++ AuthService/Authentication/AuthOperations/IAuthOperations.cs
@@ -1,7 +1,7 @@
 ﻿using AuthServiceModelLibrary.ApplicationUser;
 using AuthServiceModelLibrary.DTOs;
 
-namespace AuthService.Authentication
+namespace AuthService.Authentication.AuthOperations
 {
     public interface IAuthOperations
     {
